requirements.txt
pytest
azure-identity
azure-keyvault-secrets
databricks-cli
